+++ Assets/Packages/Plugins/UGUI/Impls/Transition/UITransitionAnimationObject.cs
@@ -10,7 +10,7 @@ namespace Cr7Sund.UGUI.Impls
     {
         public RectTransform RectTransform { get; private set; }
         public RectTransform PartnerRectTransform { get; private set; }
-        public abstract int Duration { get; }
+        public abstract long Duration { get; }
 
         void IUITransitionAnimationBehaviour.SetPartner(RectTransform partnerRectTransform)
         {
@@ -24,7 +24,7 @@ namespace Cr7Sund.UGUI.Impls
             SetTime(0);
         }
 
-        public abstract void SetTime(int time);
+        public abstract void SetTime(long time);
 
         public abstract void Setup();
     }
